fixed tests broken by change
